requirements.txt
# LLM & LangChain
langchain
langchain-core
langchain-openai
langchain-community
langgraph

# RAG e busca
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# Web frameworks
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn>=0.30.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0

# HTTP e SSL
httpx>=0.27.0
certifi>=2024.0.0

# Configuração e validação
python-dotenv>=1.0.0
pydantic>=2.10.0

# Observabilidade e utilitários
loguru>=0.7.2
tqdm>=4.66.0
